null character terminated string
At the time C (and the languages that it was derived from) was

developed, memory was extremely limited, so using only one byte of
overhead to store the length of a string was attractive. The only
popular alternative at that time, usually called a "Pascal string"
(though also used by early versions of BASIC), used a leading byte to
store the length of the string. This allows the string to contain NUL
and made finding the length need only one memory access (O(1) (constant)
time). However, C designer Dennis Ritchie chose to follow the convention
of NUL-termination, already established in BCPL,